requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.9
streamlit>=1.32.0
pandas>=2.0.0
matplotlib>=3.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
djangorestframework>=3.14.0 